githubへの通信が可能か更新チェック処理に追加

+++ shell_common/check_update.bash
@@ -21,27 +21,29 @@ else
     # Linuxの場合~/dotfile/の更新がないかチェックし、更新あればgit pullしてsetup.shを実行する
     if type git > /dev/null 2>&1 && [ -d $HOME/dotfiles ] ; then
       \cd $HOME/dotfiles
-      branch_name=`git rev-parse --abbrev-ref HEAD`
-      if [ ${branch_name} = 'master' -o ${branch_name} = 'before_zinit' ]; then
-        git fetch -p || { \cd - >/dev/null ; return 1; }
-        git checkout -q ${branch_name}
-        latest_rev=$(git ls-remote origin ${branch_name} | awk '{print $1}')
-        current_rev=$(git rev-parse HEAD)
-        if [ "$latest_rev" != "$current_rev" ]; then
-          # 最新じゃない場合には更新処理を行う
-          git reset --hard $(git log --pretty=format:%H | head -1)
-          git pull origin ${branch_name}
-          ./setup.sh
-          # tmuxがある環境であれば一時停止してログが確認できるようにする
-          if type tmux > /dev/null 2>&1; then
-            if test -z "$TMUX"; then # if not inside a tmux session
-              echo "Press Enter key to continue."
-              read wait # 一時停止
+      if \git ls-remote -h `\git remote -v | sed -e 's/\s\s*/ /g' | cut -d ' ' -f 2` >/dev/null 2>&1 ; then
+        branch_name=`git rev-parse --abbrev-ref HEAD`
+        if [ ${branch_name} = 'master' -o ${branch_name} = 'before_zinit' ]; then
+          git fetch -p || { \cd - >/dev/null ; return 1; }
+          git checkout -q ${branch_name}
+          latest_rev=$(git ls-remote origin ${branch_name} | awk '{print $1}')
+          current_rev=$(git rev-parse HEAD)
+          if [ "$latest_rev" != "$current_rev" ]; then
+            # 最新じゃない場合には更新処理を行う
+            git reset --hard $(git log --pretty=format:%H | head -1)
+            git pull origin ${branch_name}
+            ./setup.sh
+            # tmuxがある環境であれば一時停止してログが確認できるようにする
+            if type tmux > /dev/null 2>&1; then
+              if test -z "$TMUX"; then # if not inside a tmux session
+                echo "Press Enter key to continue."
+                read wait # 一時停止
+              fi
             fi
           fi
+        else
+          echo "dotfile not master branch. current branch is ${branch_name}."
         fi
-      else
-        echo "dotfile not master branch. current branch is ${branch_name}."
       fi
       \cd - >/dev/null
     fi
